pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "virallc"
version = "1.0.10"
authors = [
  { name="Chrispin Chaguza", email="chrispin.chaguza@gmail.com" },
]

maintainers = [
  {name = "Chrispin Chaguza", email = "chrispin.chaguza@gmail.com"},
]

description = "virallc: A Python package for consistent assignment of viral lineages"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research",
]

keywords = ["Virus","Lineage","Strain typing","Lineage classification","Genotype","Sublineage","Genome"]
dependencies = [
   "setuptools",
   "toml",
   "biopython",
   "fsspec",
   "pandas"
]

[tool.setuptools.packages]
find = {} 

[tool.setuptools]
py-modules = ["virallc.SeqLib.SeqLib","virallc"]

[project.scripts]
virallc = "virallc:main"
alnPairDist = "virallc.scripts.alnPairDist:main"
cleanFastaSeqs = "virallc.scripts.cleanFastaSeqs:main"
virallcLineageChecker = "virallc.scripts.virallcLineageChecker:main"
alnSeqCov = "virallc.scripts.alnSeqCov:main"
virallcClusters = "virallc.scripts.virallcClusters:main"
virallcClusters3L = "virallc.scripts.virallcClusters3L:main"

[project.urls]
Homepage = "https://github.com/ChrispinChaguza/virallc/"
Repository = "https://github.com/ChrispinChaguza/virallc.git"
Issues = "https://github.com/ChrispinChaguza/virallc/issues"
Documentation = "https://github.com/ChrispinChaguza/virallc/"
Website = "https://github.com/ChrispinChaguza/virallc"

